:recycle: Refactor App component into a separate file
