+++ tsconfig.json
@@ -9,7 +9,9 @@
         "skipLibCheck": true,
         "forceConsistentCasingInFileNames": true,
         "lib": ["DOM", "DOM.Iterable", "ESNext"],
-        "types": ["vite/client"]
+        "types": ["vite/client"],
+        "allowImportingTsExtensions": true,
+        "noEmit": true
     },
     "include": ["src"]
 }
